+++ compliant-reward-distribution/indexer-and-server/src/bin/crd_server.rs
@@ -15,7 +15,7 @@ use concordium_rust_sdk::{
     },
     signatures::verify_single_account_signature,
     smart_contracts::common::to_bytes,
-    v2::{AccountIdentifier, BlockIdentifier, Client},
+    v2::{AccountIdentifier, BlockIdentifier, Client, Upward},
     web3id::{
         did::Network,
         get_public_data, Challenge, CredentialProof,
@@ -514,8 +514,6 @@ async fn check_zk_proof(
         .response;
     let prover = account_info.account_address;
 
-    // Exclude `Initial` accounts from the proof verification.
-
     // This backend only supports regular accounts with exactly one credential (no multi-sig account support).
     if account_info.account_credentials.len() != 1 {
         return Err(ServerError::OnlyRegularAccounts);
@@ -524,9 +522,11 @@ async fn check_zk_proof(
         .account_credentials
         .get(&0.into())
         .ok_or(ServerError::OnlyRegularAccounts)?;
+
+    // Exclude `Initial` accounts from the proof verification.
     // `Initial` accounts were created by identity providers in the past
     // without a Pedersen commitment deployed on chain. As such we should not verify proofs on them.
-    if let AccountCredentialWithoutProofs::Initial { .. } = &credential.value {
+    if let Upward::Known(AccountCredentialWithoutProofs::Initial { .. }) = &credential.value {
         return Err(ServerError::NoCredentialCommitment);
     };