+++ trackAndTrace/sponsored-transaction-service/src/main.rs
@@ -278,14 +278,14 @@ pub async fn handle_transaction(
     {
         InvokeContractOutcome::Success(dry_run) => Ok(dry_run),
         InvokeContractOutcome::Failure(rejected_transaction) => {
+            let reason = rejected_transaction.reason.known_or_err()?;
+
             match rejected_transaction.decoded_reason {
                 Some(decoded_reason) => Err(ServerError::TransactionSimulationRejectedTransaction(
                     decoded_reason,
-                    rejected_transaction.reason,
+                    reason,
                 )),
-                None => Err(ServerError::TransactionSimulationError(
-                    rejected_transaction.reason,
-                )),
+                None => Err(ServerError::RejectedTransactionSimulationWithReason { reason }),
             }
         }
     }?;