+++ trackAndTrace/sponsored-transaction-service/src/types.rs
@@ -9,6 +9,7 @@ use concordium_rust_sdk::{
         Timestamp,
     },
     types::{smart_contracts::ExceedsParameterSize, Nonce, RejectReason, WalletAccount},
+    v2::upward::UnknownDataError,
 };
 use hex::FromHexError;
 use std::{
@@ -34,9 +35,13 @@ pub enum ServerError {
     /// The parameter exceeds the length limit.
     #[error("The parameter exceeds the length limit: {0}")]
     ParameterError(#[from] ExceedsParameterSize),
+    /// The transaction simulation rejected where the exact
+    /// reject reason is unknown.
+    #[error("Simulation of transaction rejected. The exact reject reason is unknown: {0}")]
+    RejectedTransactionSimulation(#[from] UnknownDataError),
     /// The transaction simulation returned with a contract rejection.
-    #[error("Simulation of transaction rejected in smart contract with reject reason: {0:?}.")]
-    TransactionSimulationError(RejectReason),
+    #[error("Simulation of transaction rejected in smart contract with reject reason: {reason:?}")]
+    RejectedTransactionSimulationWithReason { reason: RejectReason },
     /// The transaction simulation returned with a contract rejection and a
     /// decoded reject reason.
     #[error(
@@ -116,6 +121,13 @@ impl axum::response::IntoResponse for ServerError {
                     ),
                 )
             }
+            ServerError::RejectedTransactionSimulation(error) => {
+                tracing::error!("Simulation of transaction rejected. The exact reject reason is unknown: {error}.");
+                (
+                    StatusCode::INTERNAL_SERVER_ERROR,
+                    Json("Simulation of transaction rejected.".to_string()),
+                )
+            }
             error => {
                 tracing::debug!("Bad request: {error}.");
                 (StatusCode::BAD_REQUEST, Json(format!("{}", error)))
@@ -128,8 +140,8 @@ impl axum::response::IntoResponse for ServerError {
 // TODO: Use `#[from] RejectReason` instead when [`RejectReason`] implements
 // `std::error::Error`.
 impl From<RejectReason> for ServerError {
-    fn from(value: RejectReason) -> Self {
-        Self::TransactionSimulationError(value)
+    fn from(reason: RejectReason) -> Self {
+        Self::RejectedTransactionSimulationWithReason { reason }
     }
 }