+++ compliant-reward-distribution/indexer-and-server/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased changes
 
+## 0.1.3
+
+- Update `concordium-rust-sdk` dependency and adjust project to be forward-compatible. Block items with unknown details are skipped when indexing.
+
 ## 0.1.2
 
 -   Update rust-sdk submodule link.