+++ sponsoredTransactionsAuction/frontend/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased changes
 
+- Fix lint errors.
+
+## 1.0.4
+
 - Change the dApp to use new `signCIS3Message` function from the BrowserWallet.
 
 ## 1.0.3